refactor(desktop): extract host runtime health module
Move repo runtime health orchestration and the OpenCode runtime HTTP client into a focused host module so runtime_orchestrator stays responsible for lifecycle concerns. This also centralizes health result construction and snapshot handling without changing runtime health behavior.
